Authorization

+++ backend/routes/posts.js
@@ -35,7 +35,8 @@ router.post("",
         const post = new Post({
             title: req.body.title,
             content: req.body.content,
-            imagePath: url + "/images/" + req.file.filename
+            imagePath: url + "/images/" + req.file.filename,
+            creator: req.userData.userId
         }); //.body is vy body-parser
         post.save().then(createdPost => {
             res.status(201).json({
@@ -59,12 +60,20 @@ router.put("/:id", checkAuth, multer({ storage: storage }).single("image"), (req
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     }); //.body is vy body-parser
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-        res.status(200).json({
-            message: 'Update successful!'
-        });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if (result.nModified > 0) {
+            res.status(200).json({
+                message: 'Update successful!'
+            });
+        } else {
+            res.status(401).json({
+                message: 'Not Authorized'
+            });
+        }
+
     });
 })
 
@@ -103,9 +112,17 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id })
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
         .then(result => {
-            res.status(200).json({ message: "Post Deleted" });
+            if (result.n > 0) {
+                res.status(200).json({
+                    message: 'Post Deleted!'
+                });
+            } else {
+                res.status(401).json({
+                    message: 'Not Authorized'
+                });
+            }
         });
 });
